Add empty space to fix issue with non-draft mode
Compiling documents without draft mode enabled was inconsistent with
the output when draft mode was active: some institutions filled the
entire line (as expected given the arguments to makebox) whereas
others did not fill the entire line. More strangely, the results were
dependent on the initial character of the line -- e.g., an initial
'A' led to the issue whereas an initial 'B' did not.

The exact cause of this problem is uncertain, but including an empty
space (i.e., \hspace{0em}) resolves the issue. Perhaps some subtle
text manipulation leads to the observed result.