Do not recompile when version number unchanged
This changeset modifies the prerequisites for packages so that the
package is not recompiled unless its prerequisites are actually
modified. A level of indirection is added so that hgversion is no
longer the direct prerequisite of the package. Without this
indirection, the package is always rebuilt because hgversion is a
phony target.