Use LaTeX's internal commands for sections
This change improves the robustness of the section commands: instead
of redefining the commands from scratch, LaTeX's \@startsection
macro is used to control the basic formatting of the section header,
particularly the vertical spacing before and after the header.

Because resume sections should not be numbered, \thesection is also
redefined so no number is included in the header. The appendix is
also updated to omit the letters designating different sections.