Correct grep pattern for makeglossaries
This change corrects the pattern that conditionally executes
makeglossaries. First, the pattern is now positive (i.e., does an
ist file exist?). Second, single quotes are used so that the
backslash only requires a single backslash as the escape. (Using
double quotes requires four backslashes in the pattern to match a
single backslash.)