Do not ignore BibTeX or makeglossaries errors
This change does not ignore errors when running BibTeX and
makeglossaries. A non-zero exit status from grep indicates that
BibTeX or makeglossaries need not be run; hence, a minus sign
prefixes the existing rule so that the exit status from grep is
ignored. Unfortunately, this syntax also ignores the exit status from
BibTeX and makeglossaries, but these errors should cause make to stop
execution.

This change uses the shell built-in "if" command to prevent grep's
non-zero exit status from being interpreted as an error. BibTeX and
makeglossaries are conditionally executed if the grep command
succeeds.