Support variable locations of `echo`
This change provides support for non-standard locations of the echo
command. On Linux systems, GNU echo is in /bin/echo, but OS X may
have echo installed in another location (e.g., by MacPorts or
Homebrew). Instead of hard-coding the path to the executable, use
`which` to locate the version that isn't built-in to Bash.