Do not remove *.aux files in subdirectories
BibTeX fails when the aux files of included files aren't available.
This happens when the document contains an \includeonly command but
the aux files of the non-included files do not exist.

This change does not recursively search for all aux files as part of
the 'veryclean' recipe. Hence, existing aux files will be retained,
which allows BibTeX to run successfully even when an \includeonly
command is present.