Add custom index style for recipe books
This change customizes the index by bolding the main entries (i.e.,
categories) and not indenting the subentries (i.e., recipes). The file
recipe.ist provides this customization, and the Makefile is updated
to use it. The net result is a small reduction in the number of pages
required for the index.

See the manual entry for `makeindex' for more information regarding
the contents of the style file (.ist). The manual entry is also
available online -- e.g., the manual entry for Ubuntu 14.04 LTS
  http://manpages.ubuntu.com/manpages/trusty/en/man1/makeindex.1.html

References T43