Update colors per JHU's official color palette
This change updates the presentation colors to use the colors defined
in JHU's official color palette (see http://identity.jhu.edu for more
details).

This primary rationale for this change was that alerted text did not
stand out from the surrounding text due to the lack of contrast
between the dark blue (alerted) and black (normal) text. In addition
to lightening the blue color used in titles, this change also uses a
red hue for alerted text.