Add function to get the current Makefile
In many cases, it is useful to know the name of the current Makefile.
For example, if an included Makefile has a dependency on a relative
path, then it must know its current directory for that dependency.

This change adds a function that parses the current Makefile from the
MAKEFILE_LIST variable defined by GNU Make. This function may be used
to determine the current directory and for relative dependencies.

Source: "What Makefile am I in?"
(http://blog.jgc.org/2007/01/what-makefile-am-i-in.html)