Remove source code snippets from package
The source code snippets that are part of the documentation should not
appear in the actual package (i.e., generated style file). The
VerbatimOut environment is used to save each snippet in a file, and
that file is loaded via the inputminted macro.

Without this change, the source code appeared verbatim in the package,
which caused errors when the package was loaded.