Make recipe separator configurable
This changeset adds four new macros to the recipe package: asterism,
dinkus, separator, and spacebreak. The spacebreak macro will
eventually be used to separate recipes (cf. T66).

The separator macro defines the default symbols (or text) that is used
to create a visual break via the spacebreak break. The asterism and
dinkus macros are two common symbols used for this purpose; they are
included as examples and will eventually replace the "doodle" that
currently separates recipes.

Closes T65 as resolved