Vertically center logos in vbox
This changeset modifies the logo outer theme to vertically center the
institutional logo in a vertical box. Vertically centering the logo is
expected given its location in the header.

Although this change could be forced on users of this theme, but doing
so would require users to rely on internal implementation details (the
TeX register used to calculate the height of the header).

This changeset also updates the jhu.edu, engineering.jhu.edu, and
jhuapl.edu presentation themes to handle the modification to the logo
outer theme.