Support project-local texmf tree
This changeset provides experimental support for a project-local texmf
tree for all packages. The basic idea is that all class and style
files will be put in the local texmf tree instead of scattering (and
possibly duplicating) them in a myriad of project directories. The
base Makefile is updated to set the TEXINPUTS environment variable so
that TeX can find the files for \input and \openin commands.

Using a local texmf tree avoids warnings caused by the package being
in a different directory than the main file. For example,
    LaTeX Warning: You have requested package `../slides',
                   but the package provides `slides'.

Unfortunately, using TEXINPUTS to identify the local texmf tree does
not work in some cases. In particular, the epstopdf package cannot
automatically convert EPS images to PDF files when the EPS images are
in the local texmf tree. Several workarounds exist including
* setting the graphics search path (which must be done manually when
  including a package that itself includes graphics),
* writing a new wrapper for the epstopdf package that uses `kpsewhich`
  to find included EPS images (for more details, see
  http://goo.gl/bXpzQQ), and
* manually converting any EPS images used by packages to PDF files and
  storing those files in the local texmf tree.
The last workaround is the most practical but requires manual
interaction from the user and version controlling the resulting PDFs
instead of the original EPS images. (PDF is a binary format whereas
EPS images are text files.)