Search for citations in all aux files
This changeset updates the default Makefile to search for citations in
all files included by the target. That is, even if the main file does
not include any citations but included files do, then BibTeX will be
invoked. (Previously, if the main file did not include any citations,
then BibTeX would not be invoked.)

This changeset does not update the recipes for packages because it is
less likely that packages will include files.