Do not use xargs to pass arguments to grep
xargs exit status masks the exit status from grep, which means that
BibTeX is executed regardless of the presence (or absence) of citation
commands in the aux file(s). This changeset replaces xargs with the
assignment of a local variable, which is expanded to create the
complete list of auxiliary files for grep to search.