Discourage page break prior to answer
This changeset adds \nopagebreak prior to the answer in an effort to
keep the question and answer on the same page.

The current implementation of the answer environment does not allow it
to be split across pages because it is a single box. However, it would
be advantageous to split that box to allow long answers (e.g., those
with multiple inches of space) to start on one page and finish on the
next.