Support multiple optional arguments for answers
This changeset uses the pgfkeys package to handle multiple optional
arguments in answers -- e.g., the answer height or its width.

This change to the exam package is not backward compatible. The
key-value syntax must be used to specify the height -- i.e.,
    \begin{answer}[
        height=1in,
    ]
replaces the existing invocations of
    \begin{answer}[1in]
The pgfkeys package could support this conversion implicitly (e.g.,
assuming that the height is being specified and logging a deprecation
warning), but the additional effort does not seem necessary given that
a stable version of the package has yet to be released.