Fill available space with \exam@answerfill
This changeset corrects a fault introduced in the prior changeset
where \exam@answerfill did not fill all available space with dotted
lines due to saving the height of the answer in \exam@answer@height
rather than \answerheight. The correct length is now used as the
default parameter for \exam@answerfill.