Remove colors from logo theme
This changeset removes the use of "PMS Gray 4" from the author and
title header and footer lines from the logo outer theme. This color is
not defined by the theme (it appears in the bluejay color theme), and
the logo theme should not specify the colors of elements. Instead,
these colors are moved to the jhu.edu theme (which requires the
bluejay color theme).

As a side effect of this change, JHU's divisional themes are updated
to use the jhu.edu theme rather than duplicating the code that sets
the colors of the author and title in the header and footer.