Add example of seas.virginia.edu Beamer theme
This changeset adds two images of a presentation styled with the
seas.virginia.edu theme to the theme's documentation. It makes is
easier to see the immediate effect of changes to the theme.

Supporting this change is an updated default Makefile (Makefile.mk)
that uses a "canned recipe" to compile documents from documented LaTeX
source files (*.dtx) and LaTeX source files (*.tex). Unifying both
prevents duplication in the respective recipes.

The example presentation reuses the same example presentation as the
themes shown in Beamer's documentation. beamerthemeexamplebase.tex is
copied (unmodified) from Beamer's official repository
(https://goo.gl/Psoxm8), changeset 3dd33d7.

The Makefile for the seas.virginia.edu theme is updated to build the
package (i.e., style file), example presentation using the theme, and
the package documentation in that order.

Other packages (e.g., Beamer themes) used in the example have symbolic
links in the texmf directory. Note that the symbolic link, by itself,
does not guarantee that those packages are up-to-date -- i.e., it is
necessary to rebuild those packages when their prerequisites change
(e.g., when merging with another branch).