Add `tree' option to logo outer theme
This changeset adds the `tree' option to the logo outer theme. When
this option is specified, the logo theme uses the `tree' outer theme
as its base instead of the `split' outer theme.

Although this changeset bumps the version number of the logo theme,
this change should be backward compatible with the prior version.

Because of the limited number of colors specified by the various color
themes (e.g., `bluejay' and `wahoo'), additional work is required to
use this functionality with other themes (e.g., `jhu.edu').