Correct frame numbers in presentation appendix
This changeset patches the \appendix command in presentations so that
the frame number is correct for frames inserted at the beginning of
the appendix. The best example is when the "outline" option is
specified to the slides package; that option inserts the name of the
part (i.e., "Appendix") at the beginning of the appendix, but
previously, the frame number was the last of the main presentation
rather than the start of the appendix.