Add paragraph indentation for notes in Beamer
This changeset indents paragraphs after the first in notes in Beamer.
Indenting the paragraphs may not be appropriate in all cases, but it
is certainly more readable than having no indication of paragraph
breaks.

The obvious alternative to this change is using \AtBeginNote{...} to
insert code at the beginning of each note, but this macro does not
have the desired result (i.e., changes to \parindent are ignored).