Make PDFs depend on TeX files in subdirectories
This changeset alters the dependencies of PDFs compiled from (La)TeX
documents to include (La)TeX files in subdirectories but not the
current directory. This modification allows a document to detect
changes to included files (e.g., chapters or figures) but will not
cause a change to unrelated documents in the same directory to affect
each other.