Do not start section when printing glossary
The glossaries package assumes that it's context is a document (e.g.,
a report) and printing a glossary automatically inserts appropriate
section commands (e.g., \printglossary starts a new section named
"Glossary"). In a Beamer presentation, though, it may be preferable
to insert the appropriate sections manually. Hence, this changeset
overrides the default behavior to avoid starting a new section.