Use glossaries-generated ist unless compiling dtx
The glossaries package generates an index style (*.ist) when compiling
a document. This index style should be used unless the document is
documented TeX (*.dtx) -- i.e., package documentation -- in which case
the index style file should be gglo.ist.

This changeset updates the Makefile to check the type of document
being compiled and uses the index style generated by the glossaries
package unless the document filename ends in "*.dtx".