Add hyperfix package for compatibility issues
The hyperref package tends to have a large number of conflicts with
other packages. Although many packages are intentionally designed for
compatibility with hyperref, some packages -- particularly those that
are not actively maintained -- are not. The hyperfix package provides
compatibility for such packages.

This changeset creates the hyperfix package and patches the \@bibitem
macro so the hyperref package can be used with the bibentry package.