Add macro to set frequency of class meetings
Although most classes meet based on the day of the week, in some cases
the days might vary even if the frequency is consistent---e.g.,
classes that meet every other day, independent of the day of the week.
This change adds the `setclassfrequency` macro to allow the expression
of such meeting schedules.