Sum the total number of points in the exam
This change prints the total number of points in the exam at the end
of compiling the document. It eliminates the need to manually sum the
point values for all the questions when developing the exam.

In addition, this change uses the totcount package so that the total
number of points in the exam can be printed in the document.