Correct hanging indent for true/false environment
The true/false environment uses a hanging indent so that the labels
(i.e., True and False) are not obstructed by the statement. That is,
the statement does not wrap underneath the labels even when the
statement spans multiple lines of text. Unfortunately, setting a
hanging indent does not work inside LaTeX's list environments, which
are used when numbering questions. This change uses \parshape to
produce the same effect as a hanging indent regardless of the
enclosing environment(s).