Prevent infinite recursion when listing packages
If the default Makefile (`Makefile.mk`) is in the same directory as
the Makefile that included it (i.e., a "flat" project structure), then
attempting to list all the packages (`make list`) results in infinite
recursion. This change guards against this scenario by not attempting
to list packages when `Makefile.mk` is in the same directory.