Allow citations in glossaries
When a citation appears in a glossary, an additional (pdf)latex run is
required to resolve it because the glossary doesn't "exist" until
after the `makeglossaries` run. This change follows `makeglossaries`
immediately with (pdf)latex so that citations in the glossary are
handled correctly.

The additional (pdf)latex run adds overhead when the glossary does not
contain any citations. Hence, it would be ideal if the run was
configurable (or conditional) to avoid increasing the total build
time.

Stack Exchange: https://tex.stackexchange.com/a/111417