Add initial version of anonymous package
The anonymous package facilitates anonymizing documents (e.g.,
conference papers or journal articles) to facilitate double-blind
review. In particular, it allows the anonymized and non-anonymized
versions of the text to be written in parallel, and changing a package
option is all that is required to use the anonymized (or
non-anonymized) version.