Ensure line ends after full-width box
The microtype packages interferes with line breaks following a
full-width box, such as those used to set the institution and position
in a resume. More specifically, the left protrusion of certain letters
causes the box to be slightly less than the total line width, and
subsequent material (e.g., another position or descriptive text)
appears on the same line, which creates an overfull hbox as the
subsequent material extends into the right margin.

This change modifies the implementation of the institution and
position macros. The institution is set without using an explicit box
(i.e., no \makebox) but with an explicit line break. Left protrusion
is disabled for the position using the newly-defined
\noprotrusionifhmode (provided by R Schlicht in a comment to a GitHub
issue). Material following an institution or position now appears on
the subsequent line.

References schlcht/microtype#7