Strip trailing whitespace from package
When a preamble (or postamble) is included in a package, DocStrip
prepends \MetaPrefix ("%% " by default) to each line. Thus, a blank
line in the preamble (postamble) ends with a space.

Although packages are not committed to version control in this
repository, they must be when shared (i.e., used elsewhere), and
commands like `git diff --check` warn about trailing whitespace.
Consequently, this change strips any trailing whitespace -- not
limited to just that introduced by DocStrip -- from the generated
package.

Stack Exchange: https://tex.stackexchange.com/a/470230