Avoid unnecessary `pdflatex` invocations
For simple LaTeX documents (e.g., those without any references such as
a standalone image), `pdflatex` is executed three times, which is
unnecessary. This change guards the invocation of `pdflatex` after
`bibtex` and `makeindex` to avoid the unnecessary compilations.

This change sacrifices some efficiency in the case of "complex"
documents with citations, glossaries, and indices because `pdflatex`
need not be run after every auxiliary tool (e.g., `makeindex`).
Nevertheless, the increased compilation time is less likely to be
noticed in the case of such documents.

Closes #46