minted-doc: Support minted v2.6
The 2.6 release of the minted package, among other changes, modified
the implementation of Pygments style definitions to allow arbitrary
characters. Those change removed the minted@checkstyle macro, which
caused minted-doc's patches to fail.

These patches are actually unnecessary with the 2.5 release of minted,
which fixed the issue of comments in Pygments style files appearing
verbatim in documented LaTeX (*.dtx) files (see gpoore/minted#161).

This change guards the aforementioned patches so they are only applied
to versions of minted prior to the 2.5 release.

Closes #67