reproducible: Add package for reproducible builds
Comparing documents compiled by pdfTeX, specifically the resulting
PDFs, is non-trivial, and there is no easy mechanism to even identify
if a document has changed due to embedded metadata, such as the PDF's
creation time. This package ensures that the build is *reproducible*,
providing the same output each time that the document is compiled.

Closes #70