visualization: Label x axis for Likert graphs
Reviewers have commented negatively about the axis labels for the
Likert visualization, stating that the negative percentages (e.g., to
indicate disagreement) aren't intuitive. This change adds a new style,
`explicit axis labels`, for use when a legend cannot be included: the
tick mark labels on the x axis are omitted in favor of labeling the
right and left extents (as "Disagree" and "Agree" respectively).

Regardless of whether the aforementioned style is used, the negative
and positive signs on the x axis tick mark labels are removed.