resume: Allow publication type to be omitted
This change allows the type of publication (e.g., journals vs.
conferences) to be omitted from the list of publications. This format
is designed explicitly for a resume -- rather than a curriculum vitae
(CV) -- where a list of publications, if included, need not be
organized by the type of publication.

Because the "title" of the list of publications can be omitted, it
should technically be an optional argument instead of a mandatory
argument. That change is intentionally not part of this commit to
maintain backward compatibility.